+++ lib/generators/tailwindcss/controller/controller_generator.rb
@@ -4,6 +4,7 @@ module Tailwindcss
   module Generators
     class ControllerGenerator < Erb::Generators::ControllerGenerator
       source_root File.expand_path("templates", __dir__)
+      source_paths << "lib/templates/erb/controller"
     end
   end
 end