+++ lib/generators/tailwindcss/mailer/mailer_generator.rb
@@ -4,6 +4,7 @@ module Tailwindcss
   module Generators
     class MailerGenerator < Erb::Generators::MailerGenerator
       source_root File.expand_path("templates", __dir__)
+      source_paths << "lib/templates/erb/mailer"
     end
   end
 end